rename image_processor.py to processor.py
