+++ core/src/image/processor.py
